+++ scripts/embedding/yap/yap_logtalk_pl.sh
@@ -6,7 +6,7 @@
 ##   and runtime and optionally an application.pl file with a Logtalk
 ##   application
 ## 
-##   Last updated on November 1, 2024
+##   Last updated on January 14, 2025
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   SPDX-FileCopyrightText: 1998-2025 Paulo Moura <pmoura@logtalk.org>
@@ -296,7 +296,15 @@ if [ "$loader" != "" ] ; then
 	mkdir -p "$temporary/application"
 	cd "$temporary/application" || exit 1
 	yap -g "consult('$directory/logtalk'),set_logtalk_flag(clean,off),set_logtalk_flag(scratch_directory,'$temporary/application'),logtalk_load('$loader'),halt"
-	cat "$(ls -rt ./*.yap)" > "$directory"/application.pl
+	if test -n "$(find . -maxdepth 1 -name '*.yap' -print -quit)" ; then
+		files="$(ls -rt ./*.yap)"
+		for a in $files ; do cat "$a" >> "$directory"/application.pl ; done
+	else
+		echo
+		echo "No application files found!"
+		echo
+		exit 1
+	fi
 fi
 
 if [ "$saved_state" == "true" ] ; then